pipeline/cloud/services/scraper/requirements.txt
selenium
webdriver-manager
google-cloud-storage
google-cloud-logging
